ch-4/pyproject.toml
[project]
name = "ch-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.2.1",
    "email-validator>=2.2.0",
    "fastapi[all]>=0.115.8",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pyotp>=2.9.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.37",
]
